:sparkles: [open-zaak/open-zaak#1649] Command to document envvars
`decouple.config` is overridden to accept help text and several other attributes, as well as to add the loaded environment variables to a registry. This registry is used by the `generate_envvar_docs` management command to generate RST style documentation based on a template. This removes the need for duplication for these envvar docs in all the components that use open-api-framework