[#1784] fix runtime configuration for django-log-outgoing-requests
Previously, setting the `LOG_REQUESTS` setting to `False` caused the
runtime configuration of django-log-outgoing-requests to break. This was
caused by no handlers being configured for its logger.

As the runtime configuration at the moment only allows configuring
saving to the database (or not) the `save_outgoing_requests` should
always be configured in the `LOGGING` setting.